Refactor: 랭킹 재설정하는 부분 utils로 이동
