+++ controllers/bossRaid.controller.js
@@ -1,8 +1,9 @@
 const BossRaidService = require("../services/bossRaid.service");
 const RankingInfo = require("../models/rankingInfo.model");
 const getStaticData = require("../utils/getStaticData");
-require("date-utils");
+const topRankerToCache = require("../utils/topRankerToCache");
 const response = require("../utils/response");
+require("date-utils");
 
 class BossRaidController {
   static async bossRaidStatus(req, res) {
@@ -117,7 +118,7 @@ class BossRaidController {
         await BossRaidService.delRedisStatus();
   
         // 랭킹 업데이트
-        await BossRaidController.topRankerToCache();
+        await topRankerToCache();
 
         res.status(200).json({
           message: response.GAME_END,
@@ -159,22 +160,6 @@ class BossRaidController {
         topRankerInfoList: rankingInfoData,
         myRankingInfo: myRankingInfoData[0],
       });
-    } catch (err) {
-      console.log(err);
-      res.status(500).json(response.INTERNAL_SERVER_ERROR);
-    }
-  };
-
-  static async topRankerToCache(req, res) {
-    /**
-     * 기능: mysql에서 받아온 TopRanker를 캐시에 설정
-     * 작성자: 허정연
-     */
-    let rankingInfoData = [];
-    try {
-      rankingInfoData = await BossRaidService.topRankerInfoListSelect();
-      await BossRaidService.setTopRankerToCache(rankingInfoData[0]);
-      console.log(response.RANKING_RESET);
     } catch (err) {
       console.log(err);
       res.status(500).json(response.INTERNAL_SERVER_ERROR);
